services/device-selection/pyproject.toml
[project]
name = "device-selection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "outlines>=1.2.12",
    "protobuf>=6.33.6",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymoo>=0.6.1.6",
    "structlog>=25.5.0",
    "temporalio>=1.27.1",
    "tomli>=2.4.0",
    "typer>=0.24.1",
]

[project.scripts]
device-selection = "device_selection.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/device_selection"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "grpcio-tools>=1.80.0",
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
]
